cleaned up a bit persistence adapters
