+++ e2e/data-factory/persistence/entity-registry.ts
@@ -1,31 +1,30 @@
-import type {EntityRegistry, EntityMetadata} from './types';
+export interface EntityMetadata {
+    tableName?: string; // For database adapters
+    endpoint?: string; // For API adapters
+    primaryKey?: string; // Default: 'id'
+}
 
-/**
- * Default implementation of entity registry with sensible defaults
- */
+export interface EntityRegistry {
+    register(entityType: string, metadata: EntityMetadata): void;
+    getMetadata(entityType: string): EntityMetadata;
+}
+
 export class DefaultEntityRegistry implements EntityRegistry {
     private metadata = new Map<string, EntityMetadata>();
-    
+
     register(entityType: string, metadata: EntityMetadata): void {
         this.metadata.set(entityType, metadata);
     }
-    
+
     getMetadata(entityType: string): EntityMetadata {
         const meta = this.metadata.get(entityType);
-        if (!meta) {
-            // Sensible defaults - entity type maps to table name
-            return {
-                tableName: entityType,
-                primaryKey: 'id'
-            };
-        }
-        return meta;
+        return (meta === undefined) ? this.defaultMetadata(entityType) : meta;
     }
-    
-    // Helper method to register multiple entities at once
-    registerMany(entities: Record<string, EntityMetadata>): void {
-        Object.entries(entities).forEach(([entityType, metadata]) => {
-            this.register(entityType, metadata);
-        });
-    }
+
+    private defaultMetadata(entityType: string):EntityMetadata {
+        return {
+            tableName: entityType,
+            primaryKey: 'id'
+        };
+    };
 }
