+++ e2e/data-factory/persistence/index.ts
@@ -1,3 +1,4 @@
-export * from './types';
-export * from './knex-adapter';
+export * from './adapters/knex';
+export * from './adapters/tinybird';
+export * from './adapter';
 export * from './entity-registry';
