+++ e2e/data-factory/plugins/ghost/ghost-plugin.ts
@@ -3,9 +3,7 @@ import {BasePlugin} from '../base-plugin';
 import {createDatabase} from './database';
 import {PostFactory} from './posts/post-factory';
 import type {PostOptions, PostResult} from './posts/types';
-import {KnexPersistenceAdapter} from '../../persistence/knex-adapter';
-import {DefaultEntityRegistry} from '../../persistence/entity-registry';
-import type {PersistenceAdapter} from '../../persistence/types';
+import {KnexPersistenceAdapter, PersistenceAdapter, DefaultEntityRegistry} from '../../persistence';
 
 export interface GhostPluginOptions {
     database?: Knex;
@@ -18,27 +16,27 @@ export interface GhostPluginOptions {
  */
 export class GhostPlugin extends BasePlugin {
     readonly name = 'ghost';
-    
+
     private db: Knex;
     postFactory: PostFactory;
-    
+
     constructor(options: GhostPluginOptions = {}) {
         super();
-        
+
         // Set up database connection
         this.db = options.database ?? createDatabase();
-        
+
         // Set up persistence (use provided or create default)
         if (options.persistence) {
             this.setPersistenceAdapter(options.persistence);
         } else {
             this.setPersistenceAdapter(this.createDefaultPersistence());
         }
-        
+
         // Create and register factories
         this.postFactory = this.registerFactory(new PostFactory());
     }
-    
+
     /**
      * Creates the default persistence adapter for Ghost
      */
@@ -47,36 +45,36 @@ export class GhostPlugin extends BasePlugin {
         registry.register('posts', {
             tableName: 'posts'
         });
-        
+
         return new KnexPersistenceAdapter(this.db, registry);
     }
-    
+
     async setup(): Promise<void> {
         // Setup all factories
         for (const factory of this.factories.values()) {
             await factory.setup();
         }
     }
-    
+
     /**
      * Get the shared database connection for custom queries
      */
     getDatabase(): Knex {
         return this.db;
     }
-    
+
     // Allows for ghost.posts.create() syntax
     get posts(): PostFactory {
         return this.postFactory;
     }
-    
+
     // *
     // Convenience methods
     // *
     async createPost(options?: PostOptions): Promise<PostResult> {
         return this.postFactory.create(options);
     }
-    
+
     async createPublishedPost(options?: PostOptions): Promise<PostResult> {
         return this.postFactory.create({
             ...options,
@@ -84,7 +82,7 @@ export class GhostPlugin extends BasePlugin {
             published_at: options?.published_at || new Date()
         });
     }
-    
+
     async createDraftPost(options?: PostOptions): Promise<PostResult> {
         return this.postFactory.create({
             ...options,
@@ -92,7 +90,7 @@ export class GhostPlugin extends BasePlugin {
             published_at: null
         });
     }
-    
+
     async createScheduledPost(publishDate: Date, options?: PostOptions): Promise<PostResult> {
         return this.postFactory.create({
             ...options,
@@ -100,22 +98,22 @@ export class GhostPlugin extends BasePlugin {
             published_at: publishDate
         });
     }
-    
+
     /**
      * Mixed composition methods - these will coordinate multiple factories
      * once we have tags, users, etc.
      */
-    
+
     // Future: When we have tags
     // async createPostWithTags(
-    //     postOptions?: PostOptions, 
+    //     postOptions?: PostOptions,
     //     tagNames: string[] = ['default-tag']
     // ): Promise<{post: PostResult; tags: TagResult[]}> {
     //     const post = await this.createPost(postOptions);
     //     const tags = await Promise.all(
     //         tagNames.map(name => this.tagFactory.create({name}))
     //     );
-    //     
+    //
     //     // Create relationships
     //     await this.db('posts_tags').insert(
     //         tags.map(tag => ({
@@ -123,10 +121,10 @@ export class GhostPlugin extends BasePlugin {
     //             tag_id: tag.id
     //         }))
     //     );
-    //     
+    //
     //     return {post, tags};
     // }
-    
+
     // Future: When we have users/authors
     // async createAuthorWithPosts(
     //     authorOptions?: UserOptions,
@@ -134,14 +132,14 @@ export class GhostPlugin extends BasePlugin {
     // ): Promise<{author: UserResult; posts: PostResult[]}> {
     //     const author = await this.userFactory.create(authorOptions);
     //     const posts = await Promise.all(
-    //         Array(postCount).fill(null).map(() => 
+    //         Array(postCount).fill(null).map(() =>
     //             this.createPost({author_id: author.id})
     //         )
     //     );
-    //     
+    //
     //     return {author, posts};
     // }
-    
+
     /**
      * Get statistics about created entities
      */
@@ -151,14 +149,14 @@ export class GhostPlugin extends BasePlugin {
             posts: baseStats.post || 0
         };
     }
-    
+
     /**
      * Clean up and close database connection
      */
     async destroy(): Promise<void> {
         // Clean up all factories
         await super.destroy();
-        
+
         // Close database connection
         await this.db.destroy();
     }
